feat: moved reskin logic behind env variable
