+++ src/index-v2.scss
@@ -5,4 +5,4 @@
 @import "~@edx/frontend-component-header/dist/index";
 @import "~@edx/frontend-component-footer/dist/footer";
 
-@import './profile/index';
+@import './profile-v2/index';
